pyproject.toml
[tool.poetry]
name = "psqlsync"
version = "0.1.7"
description = "Tool to create basic PostgreSQL backups and restore them from local files."
authors = ["Tip ten Brink <75669206+tiptenbrink@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
psycopg2-binary = "^2.9.2"
tomli = "^2.0.1"
dirgh = { git = "https://github.com/tiptenbrink/dirgh.git", branch = "main", optional = true }
trio = { version = ">= 0.20.0, < 0.22.0", optional = true }

[tool.poetry.scripts]
psqlsync = "psqlsync.cli:run"
prep = "psqlsync.ghremote.cli:run"

[tool.poetry.extras]
dirgh = ["dirgh", "trio"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
